+++ NEWS.rst
@@ -49,6 +49,8 @@ NEWS for system-image updater
    Salveti.  (LP: #1419027)
  * A new key, ``target_version_detail`` has been added to the dictionary
    returned by the ``.Information()`` D-Bus method.  (LP: #1399687)
+ * The ``User-Agent`` HTTP header now also includes device and channel names.
+   (LP: #1387719)
 
 2.5.1 (2014-10-21)
 ==================